hiding the new listing column if there are no custom user types
