+++ src/Twig/PageBackendExtension.php
@@ -5,8 +5,7 @@ namespace OHMedia\PageBundle\Twig;
 use OHMedia\PageBundle\Entity\PageContentRow;
 use OHMedia\PageBundle\Entity\PageRevision;
 use OHMedia\PageBundle\Form\Type\PageContentRowType;
-use OHMedia\PageBundle\Util\ReadableUserType;
-use Symfony\Component\Form\FormFactoryInterface;
+use OHMedia\PageBundle\Service\PageUserTypes;
 use Twig\Environment;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
@@ -15,7 +14,7 @@ class PageBackendExtension extends AbstractExtension
 {
     private bool $rendered = false;
 
-    public function __construct(private FormFactoryInterface $formFactory)
+    public function __construct(private PageUserTypes $pageUserTypes)
     {
     }
 
@@ -30,6 +29,7 @@ class PageBackendExtension extends AbstractExtension
                 'is_safe' => ['html'],
                 'needs_environment' => 'true',
             ]),
+            new TwigFunction('has_page_user_types', [$this, 'hasUserTypes']),
             new TwigFunction('page_user_type', [$this, 'userType']),
         ];
     }
@@ -56,8 +56,13 @@ class PageBackendExtension extends AbstractExtension
         ]);
     }
 
+    public function hasUserTypes(): bool
+    {
+        return count($this->pageUserTypes->getUserTypes()) > 0;
+    }
+
     public function userType(string $type): string
     {
-        return ReadableUserType::get($type);
+        return PageUserTypes::readable($type);
     }
 }
